Fix bug wcausing 11 turns
With the previous logic, there was a bug where because a player
couldn't go twice in a row, if they were the only player with 2
turns left the game ended because there was no next valid player.

This commit fixes that bug by forcing each turn to alternate between
teams.